Fix test cases. Modify module options and add interface. Update \Service\User to verify email address

+++ src/DxUser/Service/User.php
@@ -8,9 +8,21 @@ use Zend\ServiceManager\ServiceManagerAwareInterface;
 use Zend\ServiceManager\ServiceManager;
 use Zend\Stdlib\Hydrator\ClassMethods;
 use Dxapp\EventManager\EventProvider;
+use DxUser\Options\ModuleOptions;
+use Zend\Mail\Message;
+use Zend\Mail\Transport\Sendmail as SendmailTransport;
+use Zend\View\Model\ViewModel;
+use Zend\View\Renderer\PhpRenderer;
+use DxUser\Entity\UserCodesInterface;
 
 class User extends EventProvider implements ServiceManagerAwareInterface
 {
+
+	/**
+	 * The View Renderer
+	 * @var object
+	 */
+	protected $renderer = NULL;
 
 	/**
 	 * The Doctrine Entity Manager
@@ -22,31 +34,66 @@ class User extends EventProvider implements ServiceManagerAwareInterface
 	 * the Module Options
 	 * @var type 
 	 */
-	protected $moduleOptions = NULL;
+	protected $options = NULL;
 
 	/**
 	 * Two Stage SignUp
 	 *
 	 * @param array $data The User Object
-	 * @return \DxUser\Entity\UserCodes
+	 * @return \DxUser\Entity\UserCodes|NULL
 	 * @throws Exception\InvalidArgumentException
 	 */
 	public function register($user)
 	{
-		$userCode = new \DxUser\Entity\UserCodes();
-		$userCode->setUser($user);
-		$userCode->setTypeOf('verify-email');
-		$userCode->setCode(md5($userCode->getTypeOf() . $user->getEmail()));
-		$this->getEventManager()->trigger(__FUNCTION__, $this, array('user' => $user, 'userCode' => $userCode));
-		$this->em->persist($userCode);
-		$this->getEventManager()->trigger(__FUNCTION__ . '.post', $this, array('user' => $user, 'userCode' => $userCode));
-		return $userCode;
+		if ($this->getOptions()->getEnableEmailVerification())
+		{
+			$userCode = new \DxUser\Entity\UserCodes();
+			$userCode->setUser($user);
+			$userCode->setTypeOf('verify-email');
+			$userCode->setCode(md5(time() . $userCode->getTypeOf() . $user->getEmail()));
+			$this->getEventManager()->trigger(__FUNCTION__, $this, array('user' => $user, 'userCode' => $userCode));
+			$this->em->persist($userCode);
+			$this->em->flush();
+			$this->sendVerifyEmail($userCode);
+			$this->getEventManager()->trigger(__FUNCTION__ . '.post', $this, array('user' => $user, 'userCode' => $userCode));
+			return $userCode;
+		}
+		else
+		{
+			return NULL;
+		}
 	}
 
 	/**
-	 * Verify an email address
+	 * Send email
+	 * @param object $userCode 
+	 */
+	public function sendVerifyEmail(UserCodesInterface $userCode)
+	{
+		$message = new Message();
+		$message->addFrom($this->getOptions()->getEmailVerifySender(), $this->getOptions()->getEmailVerifySender())
+				->addTo($userCode->getUser()->getEmail())
+				->setSubject($this->getOptions()->getEmailVerifySubject());
+		$viewModel = new ViewModel(array(
+					'userCode' => $userCode,
+				));
+		$viewModel->setTemplate($this->getOptions()->getTemplateVerifyEmail());
+		$body = $this->renderer->render($viewModel);
+		$message->setBody($body);
+		if ($this->getOptions()->getEmailSending())
+		{
+			$transport = new SendmailTransport();
+			$transport->send($message);
+		}
+	}
+
+	/**
+	 * Check user email verification
+	 * And verify user if needed.
 	 * @param array $data 
 	 * @throws Exception\InvalidArgumentException
+	 * 
+	 * @return boolean|string FALSE if error, TRUE if successfull, string if already verified and others.
 	 */
 	public function verifyEmail(array $data)
 	{
@@ -55,17 +102,24 @@ class User extends EventProvider implements ServiceManagerAwareInterface
 			$user = $data['user'];
 			$code = $data['code'];
 			$typeOf = 'verify-email';
-			$userCodesRepo = $this->getEntityManager()->getRepository('DxUser\Entity\UserCode');
-			$userCode = $userCodesRepo->findUserCode($user, $typeOf, $code);
-			if ($userCode)
+			if (!$user->isEmailVerified())
 			{
-				$this->getEventManager()->trigger(__FUNCTION__, $this, array('user' => $user, 'userCode' => $userCode));
-				$user->verifyEmailAddress();
-				$this->em->remove($userCode);
-				$this->em->persist($user);
-				$this->em->flush();
-				$this->getEventManager()->trigger(__FUNCTION__ . '.post', $this, array('user' => $user, 'userCode' => $userCode));
-				return TRUE;
+				$userCodesRepo = $this->getEntityManager()->getRepository('DxUser\Entity\UserCode');
+				$userCode = $userCodesRepo->findUserCode($user, $typeOf, $code);
+				if ($userCode)
+				{
+					$this->getEventManager()->trigger(__FUNCTION__, $this, array('user' => $user, 'userCode' => $userCode));
+					$user->verifyEmailAddress();
+					$this->em->persist($user);
+					$this->em->remove($userCode);
+					$this->em->flush();
+					$this->getEventManager()->trigger(__FUNCTION__ . '.post', $this, array('user' => $user, 'userCode' => $userCode));
+					return TRUE;
+				}
+			}
+			else
+			{
+				return 'already-verified';
 			}
 		}
 		return FALSE;
@@ -92,23 +146,27 @@ class User extends EventProvider implements ServiceManagerAwareInterface
 	}
 
 	/**
-	 * Set the Module Options
-	 * @param array $options
-	 * @return \DxUser\Service\User 
+	 * Get the Module OPtions from SM
+	 *
+	 * @return UserServiceOptionsInterface
 	 */
-	public function setModuleOptions($options)
+	public function getOptions()
 	{
-		$this->moduleOptions = $options;
-		return $this;
+		if (!$this->options instanceof ModuleOptions)
+		{
+			$this->setOptions($this->getServiceManager()->get('dxuser_module_options'));
+		}
+		return $this->options;
 	}
 
 	/**
-	 * Get the Module Options
-	 * @return \DxUser\Service\User 
+	 * set service options
+	 *
+	 * @param UserServiceOptionsInterface $options
 	 */
-	public function getModuleOptions()
+	public function setOptions(ModuleOptions $options)
 	{
-		return $this->moduleOptions;
+		$this->options = $options;
 	}
 
 	/**
@@ -130,6 +188,17 @@ class User extends EventProvider implements ServiceManagerAwareInterface
 	public function setServiceManager(ServiceManager $serviceManager)
 	{
 		$this->serviceManager = $serviceManager;
+		return $this;
+	}
+
+	/**
+	 * Set the ViewRenderer Object
+	 * @param type $viewRenderer
+	 * @return \DxUser\Service\User 
+	 */
+	public function setViewRenderer($viewRenderer)
+	{
+		$this->renderer = $viewRenderer;
 		return $this;
 	}
